refactor(app): 引入应用包并迁移内部入口
Why:
- 将 `app_context` 与 `facade` 的主实现迁入新的 `app/` 包，纠正上一轮仅做转发层的方向偏差
- 先让仓库内部默认走 `app.*`，为后续继续收口兼容层和内部边界打基础

What:
- 新增 `lol_audio_unpack.app` 包，并将上下文与门面实现迁入 `app/context.py`、`app/facade.py`
- 将旧 `app_context.py`、`facade.py` 改为兼容层，并把仓库内部调用切到新的 `app.*` 入口
- 补充 `app` 包入口 smoke test，并同步修正会 monkeypatch 实现模块的测试路径

Risk:
- 兼容性影响：有，仓库内部默认入口已切到 `app.*`，但旧 `app_context.py` / `facade.py` 仍保留兼容导出
- 潜在风险：若仍有未覆盖的外部调用依赖旧模块内部实现细节，需在后续兼容层清理前继续确认

Test:
- `uv run ruff check src/lol_audio_unpack/app src/lol_audio_unpack/app_context.py src/lol_audio_unpack/facade.py src/lol_audio_unpack/__init__.py src/lol_audio_unpack/cli src/lol_audio_unpack/gui src/lol_audio_unpack/manager src/lol_audio_unpack/mapping src/lol_audio_unpack/unpack src/lol_audio_unpack/runtime/remote/preparer.py src/lol_audio_unpack/model.py scripts/benchmark_cli.py scripts/benchmark_remote_live.py tests/conftest.py tests/test_app_package_api.py tests/test_app_context.py tests/test_champion_visibility.py tests/test_remote_preparer.py tests/test_cli.py tests/test_cli_invocation.py tests/test_gui_config.py tests/test_shared_data_controller.py tests/test_task_runner.py tests/test_task_creation_card.py tests/test_context_injection.py`
- `uv run pytest -q tests/test_app_package_api.py tests/test_app_context.py tests/test_context_injection.py tests/test_champion_visibility.py tests/test_remote_preparer.py tests/test_cli.py tests/test_cli_invocation.py tests/test_gui_config.py tests/test_shared_data_controller.py tests/test_task_runner.py tests/test_task_creation_card.py`
