+++ src/lol_audio_unpack/__init__.py
@@ -7,17 +7,19 @@ from pathlib import Path
 
 from loguru import logger
 
-from .app_context import (
+from .app import (
     AppConfig,
     AppContext,
     AppContextValidationError,
     AppPaths,
+    LolAudioUnpackApp,
     OperationOptions,
+    RemoteEntityCallbackPayload,
+    RemoteEntityWorkItem,
     RemoteSnapshotConfig,
     SourceMode,
     create_app_context,
 )
-from .facade import LolAudioUnpackApp, RemoteEntityCallbackPayload, RemoteEntityWorkItem
 from .manager import BinUpdater, DataReader, DataUpdater
 from .utils.logging import setup_logging
 from .utils.versioning import resolve_runtime_version
