+++ src/lol_audio_unpack/app/__init__.py
@@ -1,12 +1,11 @@
-"""应用上下文兼容层。
-
-当前首选入口已迁移到 ``lol_audio_unpack.app.context``。
-本模块仅保留旧导入路径兼容，后续阶段会删除。
-"""
+"""应用层公开入口。"""
 
 from __future__ import annotations
 
-from .app.context import (
+from .context import create_app_context
+from .facade import LolAudioUnpackApp
+from .remote import RemoteEntityCallbackPayload, RemoteEntityWorkItem
+from .types import (
     AppConfig,
     AppContext,
     AppContextValidationError,
@@ -15,7 +14,6 @@ from .app.context import (
     RemoteSnapshotConfig,
     SourceMode,
     WavOutputOptions,
-    create_app_context,
 )
 
 __all__ = [
@@ -23,7 +21,10 @@ __all__ = [
     "AppContext",
     "AppContextValidationError",
     "AppPaths",
+    "LolAudioUnpackApp",
     "OperationOptions",
+    "RemoteEntityCallbackPayload",
+    "RemoteEntityWorkItem",
     "RemoteSnapshotConfig",
     "SourceMode",
     "WavOutputOptions",
