+++ src/lol_audio_unpack/app/remote.py
@@ -1,0 +1,7 @@
+"""应用层 remote 类型入口。"""
+
+from __future__ import annotations
+
+from .facade import RemoteEntityCallbackPayload, RemoteEntityWorkItem
+
+__all__ = ["RemoteEntityCallbackPayload", "RemoteEntityWorkItem"]
