+++ src/lol_audio_unpack/cli/invocation.py
@@ -4,7 +4,7 @@ from __future__ import annotations
 
 from dataclasses import dataclass
 
-from lol_audio_unpack.app_context import OperationOptions, SourceMode, WavOutputOptions
+from lol_audio_unpack.app.context import OperationOptions, SourceMode, WavOutputOptions
 from lol_audio_unpack.config import DEFAULT_REMOTE_LIVE_REGION, DEFAULT_SHARED_SETTINGS, SettingKey
 from lol_audio_unpack.utils.runtime_paths import (
     RuntimePaths,
@@ -272,3 +272,4 @@ __all__ = [
     "render_cli_command",
     "validate_request",
 ]
+
